Add ActiveMQ Configuration, Service and Packages
The module needs to install the ActiveMQ packages from a Yum repository,
configure the middleware, then bring the service online.

For the time being, manage an exact copy of the init script shipped with
the RPM.  This script always exits with a status code of 0 when the
status method is called.

In addition, manage an exact copy of the activemq.xml configuration file
shipped with the ActiveMQ package.