Remove stages from java module
The stages feature of stdlib is turning out to be more trouble than it's worth.
Integrating the changes from Cody and Gary results in the entire mcollective
class being in the main stage.  We've already found is preferable to have
stages after main rather than before to avoid dependency cycles.  This only
works if all modules use stages.  The first module to not use stages will
create a cycle if it requires and modules who do use stages and the stages come
after main.