Add server_config class parameter
This class parameter allows the end user of the module
to supply their own configuration file contents without
modification to the activemq module.

In addition, a small change to the relationships.  Do
not require a resource in a 3rd party module, instead
require the class itself as per the style guide and do so
as close to init.pp as possible.