on debian-based systems use $instance
Debian-based systems have the instances-available thing going on, which is
different enough to the activemq of redhat to be problematic.

Move closer to being able to cope by managing just one instance, by default
called activemq.

An alternate approach might be to just take a `$config_prefix` parameter, but
then users would still need to activate the instance/create that instance
directory outside of the module.